refactor(SoC): 使用io.hpp重构断言实现
1. 使用io.hpp重构断言实现
2. 将系统定时器设为64位

+++ SoC/include/utils.hpp
@@ -206,7 +206,8 @@ namespace SoC
         ::SoC::detail::wait_for(cycles);
     }
 
-    inline constinit ::std::size_t systick{};
+    // 系统时刻数
+    inline constinit ::std::uint64_t systick{};
 }  // namespace SoC
 
 namespace SoC
